Extract loadToots() into api.ts
